GPG Modification

+++ Dockerfile
@@ -69,11 +69,21 @@ MAINTAINER Dave Conroy <dave at tiredofit dot ca>
 			geoip-dev \
 			perl-dev && \
 		curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz -o nginx.tar.gz && \
-        curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz.asc  -o nginx.tar.gz.asc && \
-        export GNUPGHOME="$(mktemp -d)" && \
-        gpg --keyserver ha.pool.sks-keyservers.net --recv-keys "$GPG_KEYS" && \
-        gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz && \
-        rm -r "$GNUPGHOME" nginx.tar.gz.asc && \
+	    curl -fSL http://nginx.org/download/nginx-$NGINX_VERSION.tar.gz.asc  -o nginx.tar.gz.asc && \ 
+	    export GNUPGHOME="$(mktemp -d)" && \
+	    found=''; \
+			for server in \
+				ha.pool.sks-keyservers.net \
+				hkp://keyserver.ubuntu.com:80 \
+				hkp://p80.pool.sks-keyservers.net:80 \
+				pgp.mit.edu \
+			; do \
+				echo "Fetching GPG key $GPG_KEYS from $server"; \
+				gpg --keyserver "$server" --keyserver-options timeout=10 --recv-keys "$GPG_KEYS" && found=yes && break; \
+		    done; \
+	    test -z "$found" && echo >&2 "error: failed to fetch GPG key $GPG_KEYS" && exit 1; \
+	    gpg --batch --verify nginx.tar.gz.asc nginx.tar.gz && \
+	    rm -r "$GNUPGHOME" nginx.tar.gz.asc && \
         mkdir -p /usr/src && \
         tar -zxC /usr/src -f nginx.tar.gz && \
         rm nginx.tar.gz && \
